Rename/move bookmark edit form file
it's turned from a full view in the popup to just a component
also rename its onSave to onUpdate to better refect the usage
